+++ WordPress/Tests/Security/NonceVerificationUnitTest.php
@@ -7,7 +7,7 @@
  * @license https://opensource.org/licenses/MIT MIT
  */
 
-namespace WordPress\Tests\CSRF;
+namespace WordPress\Tests\Security;
 
 use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
 
@@ -18,9 +18,7 @@ use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
  *
  * @since   0.5.0
  * @since   0.13.0 Class name changed: this class is now namespaced.
- * @since   0.15.0 The sniff has been deprecated. This unit test file now
- *                 only tests that the deprecation warnings are correctly thrown
- *                 and that the sniff falls through to the parent correctly.
+ * @since   0.15.0 This sniff has been moved from the `CSRF` category to the `Security` category.
  */
 class NonceVerificationUnitTest extends AbstractSniffUnitTest {
 
@@ -32,7 +30,22 @@ class NonceVerificationUnitTest extends AbstractSniffUnitTest {
 	public function getErrorList() {
 
 		return array(
-			4 => 1,
+			5   => 1,
+			9   => 1,
+			31  => 1,
+			44  => 1,
+			48  => 1,
+			69  => 1,
+			89  => 1,
+			113 => 1,
+			114 => 1,
+			122 => 1,
+			126 => 1,
+			148 => 1,
+			150 => 1,
+			159 => 1,
+			160 => 1,
+			161 => 1,
 		);
 	}
 
@@ -42,9 +55,7 @@ class NonceVerificationUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of warnings>
 	 */
 	public function getWarningList() {
-		return array(
-			1 => 2,
-		);
+		return array();
 
 	}
