+++ WordPress/ruleset.xml
@@ -81,4 +81,14 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.CSRF.NonceVerification.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.CSRF.NonceVerification.NoNonceVerification">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
